fix: keep trace aggregation audit-clean
