+++ src/core/extension/trace/aggregate_report.rs
@@ -1,0 +1,8 @@
+use serde::Serialize;
+
+#[derive(Serialize, Clone)]
+pub struct TraceAggregateSpanSampleOutput {
+    pub run_index: usize,
+    pub duration_ms: u64,
+    pub artifact_path: String,
+}
